Working potentiometer and button

+++ arduino/lib/LaTask/LaTask.h
@@ -12,7 +12,7 @@ class LaTask : public Task {
   Led* led;
   Photoresistor* photoresistor;
   Pir* pir;
-  float time = 0.0;
+  unsigned long time = 0;
 
   public:
     LaTask(Led* led, Photoresistor* photoresistor, Pir* pir);
